requirements.txt
fastapi[standard]>=0.128.6,<0.129.0
pydantic-settings>=2.10.1,<2.11.0
httpx>=0.28.1,<0.29.0
python-multipart>=0.0.22,<0.0.23
uvicorn[standard]>=0.40.0,<0.41.0
slowapi>=0.1.9,<0.2.0
langchain[standard]>=1.2.9,<1.3.0
langchain-community>=0.4.1,<0.5.0
chromadb>=1.5.0,<1.6.0
sentence-transformers>=5.2.2,<5.3.0
pypdf>=6.7.0,<6.8.0
ddgs>=9.10.0
langchain-huggingface>=1.1.0,<1.2.0
langchain-chroma
llama-cpp-python
huggingface-hub
pytest
pytest-asyncio
pytest-cov
langchain-experimental
secure